Move algo options to separate config file and document it.
